Switch engine keybinding to be a keyword argument
Previously, the key to which an engine was bound was an *optional*
argument, so we'd say something like:

(defengine duckduckgo
  "https://duckduckgo.com/?q=%s"
  "d")

This commit makes it a *keyword* argument. We'll now have to say:

(defengine duckduckgo
  "https://duckduckgo.com/?q=%s"
  :keybinding "d")

BE FOREWARNED: This is a breaking change! However, we're now able to add
additional optional keyword arguments to defengine. We've had requests
for features like:

* Specifying engine-specific browsers
* Override docstrings on a per-engine basis
* Specify character encoding for a particular engine

This should make adding per-engine features much easier.