Allow defengine to assign custom docstrings
When defengine creates a search function it assigns it a generic
automatically-generated docstring. This looks like:

"Search [engine-name] for the selected text. Prompt for input if none is
provided."

That's fine, but sometimes we want to be able to assign our own custom
docstrings. This commit provides an "docstring" keyword argument to
defengine, which we can use as follows:

(defengine ctan
  "http://www.ctan.org/search/?x=1&PORTAL=on&phrase=%s"
  :keybinding "c"
  :docstring "Search the Comprehensive TeX Archive Network (ctan.org)")

Thanks to @clemso for the suggestion and preliminary implementation!