Provide a hook to transform the search term
Sometimes, for a given engine, we want to apply a function to a search
term before it's interpolated into the URL. This commit provides a
`term-transformation-hook` keyword argument in `defengine` that takes a
function and ensures that it'll be invoked, with the raw term as its
argument, every time the engine is evaluated.

For example, suppose we want to ensure that search terms are encoded as
latin-1. We could pass in an anonymous function:

(defengine duckduckgo
  "https://duckduckgo.com/?q=%s"
  :term-transformation-hook (lambda (term) (encode-coding-string term latin-1)))

Thanks to @patxoca for the suggestion and preliminary implementation!