Add a custom variable to set the default browser
Previously, if a user wanted to change the default browser for all
engines, they had to overwrite `browse-url-browser-function`. That's not
great, since Emacs uses that globally to open *all* links.

Now, to set the default browser for engine-mode only, users can
overwrite the `engine/browser-function` variable instead.
`engine/browser-function` defaults to `browse-url-browser-function`, so
there shouldn't be any compatibility issues.

Thanks very much to @asok for the idea and initial implementation!