Nest prefixed keymaps correctly
We'd been handling prefixes by defining a string `engine/keymap-prefix`
which contained the prefix, concatenating that with the `:keybinding`
when an engine was defined, and then binding the engine to that
resulting string. That's not great:

* Sometimes perfectly valid prefix keys aren't represented as
  strings (e.g. `(kbd "s-/")`), so concatenation fails.
* Engines can't be automatically rebound to a new prefix once they've
  been defined.
* Keymaps are just cleaner and easier to handle.

Instead, we're defining a proper keymap and doing away with
`engine/keymap-prefix`. This isn't great for backwards compatibility,
since folks may have been relying on setting that to get the correct
keymap. Sorry about that. =/

Now users will need to use `engine/set-prefix-key`, which binds the
keymap. This has the advantage of being a function, instead of just a
variable, so we should be able to modify it in the future without
disturbing the interface.