Update and rename engine-mode.el to engine.el
I hacked it pretty severely.  Main thing is I got rid of the minor mode, as it didn't do anything except switch on the keymap, which is trivial to do by hand.  Consequently I renamed the module to just engine.el.  I also added the ability to default the search term to something else than the proximate symbol.  This can be useful for googling error messages, for example.  And I added a menu, which for my purposes works better than a keymap.  I use tmm.el and this frees me from the necessity to remember the keys.

These are the major changes, other than that just stylistic changes to satisfy my lust for code density.

Seeing that you just released, I don't expect you to merge this whole, I just want to have it on record if someone with tastes closer to mine comes by.