Prevent browse-url-browser-function shadowing
Engines, by default, use the built-in `browse-url-browser-function` to decide
which browser to open for a given search. However, if that function has been
locally bound to another browser function, it'll use that new function instead!
That's not the intended behavior.

For example, suppose a user defines an engine to search Wikipedia using their
default browser, then tries to invoke it from within an `eww` buffer. They'd
*expect* their default browser to open, but instead, because `eww` locally
overwrites `browse-url-browser-function`, their search will be opened in `eww`!

This explicitly binds the `engine/browser-function` to
`browse-url-browser-function`. That ensures that the default browser will be set
at macro-expansion, so future redefinitions of `browse-url-browser-function`
shouldn't affect engines.

This is technically a breaking change, since users might expect that redefining
`browse-url-browser-function` after defining engines would cause those engines
to use the new default browser. The documentation doesn't *explicitly* promise
that, though, and I tend to think this implementation is slightly less
surprising.

Users who *do* want to change the default browser on a mode-by-mode basis should
still be able to do so by defining a hook:

    (add-hook 'my-favorite-mode-hook
              (lambda () (setq engine/browser-function 'my-favorite-browsing-function)))

Fixes #40.