Allow binding engines to non-prefix key sequences
We'd like to be able to bind some engines to sequences of keys "nested" under
the common engine-mode prefix. For example, something like this ought to work:

```
(defengine wikipedia-spanish
    "https://es.wikipedia.org/w/index.php?search=%s"
    :keybinding "w s")
```

This makes that possible by replacing the use of `define-key` (which takes a key
sequence, which needs to start with a prefix key) with the recommended function
`keymap-set` (which takes any string that satisfies `key-valid-p`).

Closes #56.