Add melpazoid GitHub workflow
For linting packages in MELPA.

However, this doesn't fail the *action*, even if linting fails. `engine-mode`
has been around for a while, and renaming some functions might take some time,
so `package-lint` is likely to keep failing for the foreseeable future. I
appreciate melpazoid's advice, and I want to keep it around with the eventual
goal of enabling it "for real," but I'm not ready to implement all its
suggestions yet, so this ensures that it doesn't fail a build even if its checks
fail.

https://github.com/riscy/melpazoid