Prefix "private" functions with `engine--`
...instead of `engine/`. We're moving toward a more conventional `-` and `--`
prefix separator for functions, and renaming the "private" ones that are only
used internally (hopefully!) is a step in that direction.

This also fixes some `package-lint` violations.

Historically: when @hrs first wrote this thing back in 2014 that package prefix
separator convention was less standardized than it is now, and `/` has since
lost (sadly, I think, since by separating the package name from the function or
variable name the `/` conveys additional semantic information, and it's no
harder to type than a `-` on most keyboards, but so it goes).